Adding experimental code for reverse engineering DOSDPs, see https://github.com/dosumis/dead_simple_owl_design_patterns/issues/9
In particular this allows:

 * generation of DOSDP YAML based on concise prolog syntax (need the reverse operation too)
 * extraction of CSVs based on pattern matching the OWL equivalence clause in a DOSDP
 * extraction of CSVs based on pattern matching the lexical (name and synonym) patterns in a DOSDP (ie making obol DOSDP compliant)

While doing this in prolog has some obvious disadvantages, it was quick for me to get up and running. We can explore
a conversion to another language later. Or alternatively it should be fairly easy to make this code independent of the blipkit behemoth,
and work direcly with swi rdf libs.

For example of use, see:

 * https://github.com/obophenotype/human-phenotype-ontology/pull/1028
 * https://github.com/obophenotype/uberon/issues/1257